ai-workflow-system-complete-20251229_125309/ai-workflow-system-complete-20251229_125248/projects_review/morewritings/requirements.txt
openai>=1.0.0
click>=8.0.0
pydantic>=2.0.0
PyYAML>=6.0.0
